// src/app/models/request/patient-request.ts
export interface PatientRequest {

  name: string;
  surname: string;
  healthInsurance: string;
  genre: string;
  plan: string;
  affiliateNumber: string;
  dateOfBirth: Date;
  cellphone: string;
  landline: string;
  nationality: string;
  province: string;
  address: string;
  profession: string;
  email: string;
  dni: string;
  age: string;

}